refactor(features): consolidate row state into useFeatureRowState hook
Replace useOptimisticToggle with useFeatureRowState that handles both
toggle and remove loading states. This simplifies FeatureRow by moving
all row-level state management into a single, well-documented hook.
